+++ Controllers/Backend/Lengow.php
@@ -43,7 +43,9 @@ class Shopware_Controllers_Backend_Lengow extends Shopware_Controllers_Backend_E
                 'success' => true,
                 'data'    => array(
                     'panelHtml'     => $panelHtml,
-                    'isNewMerchant' => Shopware_Plugins_Backend_Lengow_Components_LengowMain::isNewMerchant())
+                    'isNewMerchant' => Shopware_Plugins_Backend_Lengow_Components_LengowMain::isNewMerchant(),
+                    'langIsoCode'   => substr(Shopware_Plugins_Backend_Lengow_Components_LengowMain::getLocale(), 0, 2)
+                )
             )
         );
     }